requirements.txt
Django~=3.1.2
django-environ~=0.4.5
django-cleanup~=5.1.0
django-import-export~=2.4.0
whitenoise~=5.2.0
gunicorn~=19.9.0
psycopg2-binary~=2.8.6
